ai_detector_unimodel/requirements.txt
flask==2.3.3
nltk==3.8.1
requests==2.31.0
pyspellchecker==0.7.2
opencv-python==4.8.1.78
pillow==10.0.1
numpy==1.24.3
scikit-image==0.21.0
tensorflow==2.13.0
scikit-learn==1.3.0
matplotlib==3.7.2
mediapipe==0.10.14
joblib==1.3.2
librosa==0.10.1
soundfile==0.12.1
language-tool-python
pandas